move task components into correct folder
